+++ mod.ts
@@ -18,3 +18,4 @@ export { default as Markdown } from './components/Markdown/index.tsx';
 export { default as Gradient } from './components/Gradient/index.tsx';
 export { default as Chip } from './components/Chip/index.tsx';
 export { default as Chiplist } from './components/Chiplist/index.tsx';
+export { default as Fieldset } from './components/Fieldset/index.tsx';