+++ routes/test/input.tsx
@@ -1,4 +1,12 @@
-import { Header, Input, Main, Select, Text, TextArea } from '../../mod.ts';
+import {
+  Fieldset,
+  Header,
+  Input,
+  Main,
+  Select,
+  Text,
+  TextArea,
+} from '../../mod.ts';
 
 export default function TestInput() {
   const selectData = ['a', 'b', 'c'];
@@ -27,6 +35,24 @@ export default function TestInput() {
           <Select options={selectData} label='label' />
           <Select options={selectData} error='error' label='label' />
           <Select maxWidth options={selectData} />
+
+          <Text type='subheading'>Fieldset:</Text>
+          <Fieldset
+            name='test-a'
+            legend='radio fieldset'
+            values={['option a', 'option b', 'option c']}
+          />
+          <Fieldset
+            name='test-a'
+            legend='checkbox fieldset'
+            values={['option a', 'option b', 'option c']}
+            allowMultiple
+          />
+
+          <Text type='subheading'>Other inputs</Text>
+          <Input type='range' />
+          <Input type='color' />
+          <Input type='file' />
         </>
       </Main>
     </>